fix: 엔티티 수정 - like가 Mysql 예약어이므로 클래스명을 likes로 변경 - user 클래스명을 users로 변경 - sql 문법 오류로 인해 columnDefinition 어노테이션 부분 수정
