--- src/main/java/com/efub/lakkulakku/domain/user/dto/LoginResDto.java
@@ -1,4 +1,0 @@
-package com.efub.lakkulakku.domain.user.dto;
-
-public class LoginResDto {
-}
