+++ src/main/java/com/efub/lakkulakku/domain/users/controller/LoginController.java
@@ -1,9 +1,9 @@
-package com.efub.lakkulakku.domain.user.controller;
+package com.efub.lakkulakku.domain.users.controller;
 
-import com.efub.lakkulakku.domain.user.dto.SignupReqDto;
-import com.efub.lakkulakku.domain.user.exception.DuplicateEmailException;
-import com.efub.lakkulakku.domain.user.exception.DuplicateNicknameException;
-import com.efub.lakkulakku.domain.user.service.UserService;
+import com.efub.lakkulakku.domain.users.dto.SignupReqDto;
+import com.efub.lakkulakku.domain.users.exception.DuplicateEmailException;
+import com.efub.lakkulakku.domain.users.exception.DuplicateNicknameException;
+import com.efub.lakkulakku.domain.users.service.UsersService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -12,19 +12,19 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/v1/users")
 @RequiredArgsConstructor
 public class LoginController {
-    private final UserService userService;
+    private final UsersService usersService;
 
 
     @PostMapping("/signup")
     public ResponseEntity<String> signup(@RequestBody SignupReqDto reqDto) {
-        userService.signup(reqDto);
+        usersService.signup(reqDto);
         //나중에 토큰 responseBody로 보내기
         return ResponseEntity.ok("성공적으로 가입되었습니다.");
     }
 
     @GetMapping("/signup/email")
     public ResponseEntity<?> checkEmailDuplicate(@RequestParam String email) {
-        if (userService.checkEmailDuplicate(email) == true) {
+        if (usersService.checkEmailDuplicate(email) == true) {
             throw new DuplicateEmailException();
         } else {
             return ResponseEntity.ok("사용할 수 있는 이메일입니다.");
@@ -34,7 +34,7 @@ public class LoginController {
 
     @GetMapping("/signup/nickname")
     public ResponseEntity<?> checkNicknameDuplicate(@RequestParam String nickname) {
-        if (userService.checkNicknameDuplicate(nickname) == true) {
+        if (usersService.checkNicknameDuplicate(nickname) == true) {
             throw new DuplicateNicknameException();
         } else {
             return ResponseEntity.ok("사용할 수 있는 닉네임입니다.");
