+++ src/main/java/com/efub/lakkulakku/domain/users/dto/LoginReqDto.java
@@ -1,0 +1,4 @@
+package com.efub.lakkulakku.domain.users.dto;
+
+public class LoginReqDto {
+}
