+++ src/main/java/com/efub/lakkulakku/domain/users/dto/LoginResDto.java
@@ -1,0 +1,4 @@
+package com.efub.lakkulakku.domain.users.dto;
+
+public class LoginResDto {
+}
