+++ src/main/java/com/efub/lakkulakku/domain/users/entity/Users.java
@@ -1,4 +1,4 @@
-package com.efub.lakkulakku.domain.user.entity;
+package com.efub.lakkulakku.domain.users.entity;
 
 import com.efub.lakkulakku.domain.profile.entity.Profile;
 import com.efub.lakkulakku.global.entity.BaseTimeEntity;
