+++ src/main/java/com/efub/lakkulakku/domain/users/repository/UsersRepository.java
@@ -1,13 +1,13 @@
-package com.efub.lakkulakku.domain.user.repository;
+package com.efub.lakkulakku.domain.users.repository;
 
-import com.efub.lakkulakku.domain.user.entity.Users;
+import com.efub.lakkulakku.domain.users.entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.UUID;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, UUID> {
+public interface UsersRepository extends JpaRepository<Users, UUID> {
     boolean existsByNickname(String nickname);
 
     boolean existsByEmail(String email);
