+++ src/main/resources/application.yml
@@ -4,9 +4,9 @@ spring:
       use-legacy-processing: true
       profiles: local
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: 
+    url: jdbc:mysql://localhost:3306/EFUB?serverTimezone=UTC&characterEncoding=UTF-8
     username: root
-    password: 
+    password: Kjy990216!
 
   jpa:
     hibernate:
